+++ src/main/java/com/github/henryco/injector/Container.java
@@ -1,6 +1,7 @@
 package com.github.henryco.injector;
 
 import com.github.henryco.injector.meta.ModuleStruct;
+import com.github.henryco.injector.meta.resolver.IClassFinder;
 
 import java.util.HashSet;
 import java.util.Set;
@@ -19,13 +20,14 @@ public final class Container {
 		modules = new HashSet<>();
 	}
 
-	void addModules(Class<?> ... modules) {
+	void addModules(IClassFinder classFinder, Class<?> ... modules) {
 		for (Class<?> module : modules) {
 			if (this.modules.stream().anyMatch(m -> m.getModule().getName().equals(module.getName())))
 				continue;
-			this.modules.add(new ModuleStruct(module));
+			this.modules.add(new ModuleStruct(module, classFinder));
 		}
 	}
+
 
 	void reset() {
 		modules.clear();