+++ src/main/java/com/github/henryco/injector/meta/resolver/IClassFinder.java
@@ -1,0 +1,7 @@
+package com.github.henryco.injector.meta.resolver;
+
+import java.util.List;
+
+public interface IClassFinder {
+	List<Class<?>> getClassesForPackage(String pckgname) throws ClassNotFoundException;
+}