+++ src/test/java/com/github/henryco/injector/general/component/ComponentGA.java
@@ -1,17 +1,21 @@
 package com.github.henryco.injector.general.component;
 
-import com.github.henryco.injector.meta.annotations.Component;
-import com.github.henryco.injector.meta.annotations.Inject;
+
+import com.github.henryco.injector.meta.annotations.Provide;
+
+import javax.inject.Inject;
+import javax.inject.Named;
 
 /**
  * Created 12/18/2017
  *
  * @author Henry
  */
-@Component("ComponentGA")
+@Provide("ComponentGA")
 public class ComponentGA implements IComponent {
 
-	@Inject("component_gb")
+	@Inject
+	@Named("component_gb")
 	private IComponent component;
 
 	private NotAnnotatedComponentB componentB;
@@ -33,5 +37,13 @@ public class ComponentGA implements IComponent {
 	@Inject
 	public void setComponentB(NotAnnotatedComponentB componentB) {
 		this.componentB = componentB;
+	}
+
+	@Override
+	public String toString() {
+		return "ComponentGA{" +
+				"component=" + component +
+				", componentB=" + componentB +
+				'}';
 	}
 }