+++ src/test/java/com/github/henryco/injector/injector/component/ComponentC.java
@@ -1,12 +1,14 @@
 package com.github.henryco.injector.injector.component;
 
-import com.github.henryco.injector.meta.annotations.Component;
-import com.github.henryco.injector.meta.annotations.Singleton;
+
+import com.github.henryco.injector.meta.annotations.Provide;
+
+import javax.inject.Singleton;
 
 /**
  * @author Henry on 17/12/17.
  */
-@Component
+@Provide
 @Singleton
 public class ComponentC {