+++ src/test/java/com/github/henryco/injector/injector/module/ModuleA.java
@@ -2,9 +2,10 @@ package com.github.henryco.injector.injector.module;
 
 import com.github.henryco.injector.injector.component.ComponentA;
 import com.github.henryco.injector.injector.component.ComponentD;
-import com.github.henryco.injector.meta.annotations.Inject;
 import com.github.henryco.injector.meta.annotations.Module;
 import com.github.henryco.injector.meta.annotations.Provide;
+
+import javax.inject.Named;
 
 /**
  * @author Henry on 17/12/17.
@@ -14,8 +15,7 @@ public class ModuleA {
 
 
 	@Provide("provide_D")
-	public ComponentD provideD(ComponentA componentA, @Inject("number2") int value) {
-		System.out.println("provideD(" + componentA + ", " + value +")");
+	public ComponentD provideD(ComponentA componentA, @Named("number2") int value) {
 		return new ComponentD(componentA.toString(), value + "");
 	}