+++ src/test/java/com/github/henryco/injector/injector/module/ModuleC.java
@@ -11,14 +11,12 @@ public class ModuleC {
 
 	@Provide("number1")
 	public int provideNumber1() {
-		System.out.println("provideNumber1()");
 		return 123;
 	}
 
 
 	@Provide("number2")
 	public int provideNumber2() {
-		System.out.println("provideNumber2()");
 		return 4;
 	}