+++ CHANGELOG.md
@@ -1,7 +1,22 @@
-15-01-2018
-Version 1.0.1
-Added shared modules
+## 10-06-2018
+##### Version 1.2.0 Release
+* Added JSR-330 Compatibility
 
-13-01-2018 
-Version 1.0.0
-Release
+## 22-01-2018
+##### Version 1.1.0 Release
+
+## 21-01-2018
+##### Version 1.0.3, 1.0.4, 1.0.5
+* Added CustomClassFinder interface and bug fix
+
+## 20-01-2018
+##### Version 1.0.2
+* Added @Named annotation
+
+## 15-01-2018
+##### Version 1.0.1
+* Added shared modules
+
+## 13-01-2018 
+##### Version 1.0.0
+* Release